feat!: Extract logic with registering event listeners to a separate file
